# market_service/target/generated-sources/payment-client/api/openapi.yaml
openapi: 3.0.3
info:
  title: API сервиса платежей
  version: 1.0.0
servers:
- url: /
paths:
  /balance/{sessionId}:
    get:
      operationId: getBalance
      parameters:
      - description: Идентификатор сессии
        explode: false
        in: path
        name: sessionId
        required: true
        schema:
          $ref: "#/components/schemas/SessionId"
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentValue"
          description: Сумма баланса
      summary: Получить баланс
      tags:
      - balance
      x-accepts:
      - application/json
  /payment/{sessionId}:
    post:
      operationId: payment
      parameters:
      - description: Идентификатор сессии
        explode: false
        in: path
        name: sessionId
        required: true
        schema:
          $ref: "#/components/schemas/SessionId"
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentValue"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentValue"
          description: Сумма баланса
      summary: Вычесть из баланса
      tags:
      - balance
      x-content-type: application/json
      x-accepts:
      - application/json
components:
  parameters:
    SessionId:
      description: Идентификатор сессии
      explode: false
      in: path
      name: sessionId
      required: true
      schema:
        $ref: "#/components/schemas/SessionId"
      style: simple
  responses:
    SuccessBalance:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PaymentValue"
      description: Сумма баланса
  schemas:
    PaymentValue:
      description: Денежное значение
      format: int64
      nullable: false
      type: integer
    SessionId:
      description: Идентификатор сессии
      nullable: false
      type: string