# payment_api/target/generated-sources/openapi/src/main/java/ru/ya/practicum/payment/api/BalanceApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.21.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package ru.ya.practicum.payment.api;

import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.server.ServerWebExchange;
import reactor.core.publisher.Flux;
import reactor.core.publisher.Mono;
import org.springframework.http.codec.multipart.Part;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-04-02T14:41:11.378193100+03:00[Europe/Moscow]", comments = "Generator version: 7.21.0")
@Validated
@Tag(name = "balance", description = "the balance API")
public interface BalanceApi {

    String PATH_GET_BALANCE = "/balance/{username}";
    /**
     * GET /balance/{username} : Получить баланс
     *
     * @param username Уникальное имя пользователя (required)
     * @return Сумма баланса (status code 200)
     */
    @Operation(
        operationId = "getBalance",
        summary = "Получить баланс",
        tags = { "balance" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Сумма баланса", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Long.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = BalanceApi.PATH_GET_BALANCE,
        produces = { "application/json" }
    )
    default Mono<ResponseEntity<Long>> getBalance(
        @Parameter(name = "username", description = "Уникальное имя пользователя", required = true, in = ParameterIn.PATH) @PathVariable("username") String username,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        Mono<Void> result = Mono.empty();
    

        exchange.getResponse().setStatusCode(HttpStatus.NOT_IMPLEMENTED);

        return result.then(Mono.empty());

    }


    String PATH_PAYMENT = "/payment/{username}";
    /**
     * POST /payment/{username} : Вычесть из баланса
     *
     * @param username Уникальное имя пользователя (required)
     * @param body  (required)
     * @return Сумма баланса (status code 200)
     */
    @Operation(
        operationId = "payment",
        summary = "Вычесть из баланса",
        tags = { "balance" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Сумма баланса", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Long.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = BalanceApi.PATH_PAYMENT,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default Mono<ResponseEntity<Long>> payment(
        @Parameter(name = "username", description = "Уникальное имя пользователя", required = true, in = ParameterIn.PATH) @PathVariable("username") String username,
        @Parameter(name = "body", description = "", required = true) @Valid @RequestBody Mono<Long> body,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        Mono<Void> result = Mono.empty();
    

        exchange.getResponse().setStatusCode(HttpStatus.NOT_IMPLEMENTED);

        return result.then(body).then(Mono.empty());

    }

}